Allow both point and polygon layers to display
