+++ packages/tupaia-web/src/features/Map/MapOverlays/DataVisualsLayer/index.ts
@@ -1,0 +1,1 @@
+export {DataVisualsLayer} from './DataVisualsLayer';
